Created music_album and genre classes
Create a Genre class with an association to the Item class (in a separate .rb file).
All MusicAlbum class properties visible in the diagram should be defined and set up in the constructor method.
All Genre class properties visible in the diagram should be defined and set up in the constructor method.